Rename module to knockpymetasploit3
