+++ tests/test_console.py
@@ -1,7 +1,7 @@
 import pytest
 import time
 
-from pymetasploit3.msfrpc import *
+from knockpymetasploit3.msfrpc import *
 
 @pytest.fixture()
 def client():
