+++ tests/test_db.py
@@ -2,7 +2,7 @@
 
 import pytest
 import os
-from pymetasploit3.msfrpc import *
+from knockpymetasploit3.msfrpc import *
 
 
 @pytest.fixture()
